Pakeitimai.
