+++ include/Block.h
@@ -51,7 +51,10 @@ public:
         return false;
     }
 
-    ~Block(){}
+    ~Block()
+    {
+        
+    }
 };
 
 #endif
