+++ include/Register.h
@@ -3,6 +3,8 @@
 
 #include <cstdint>
 
+#include "Word.h"
+
 class Register
 {
 private:
@@ -23,7 +25,10 @@ public:
         this->word.get_int();
     }
 
-    ~Register();
+    ~Register()
+    {
+
+    }
 };
 
 class Flag{
@@ -38,6 +43,11 @@ public:
     bool check_status()
     {
         return status;
+    }
+
+    ~Flag()
+    {
+        
     }
 }
