+++ src/RealMachine.cpp
@@ -1,0 +1,11 @@
+#include "RealMachine.h"
+
+RealMachine::RealMachine()
+{
+    
+}
+
+RealMachine::~RealMachine()
+{
+
+}
